refactor(arch): move fetch_vm_info to detection_service and fix test assertions
