+++ src/osx_proxmox_next/planner.py
@@ -332,38 +332,7 @@ def _apple_services_steps(config: VmConfig, vmid: str) -> list[PlanStep]:
 
 # ── VM Destroy ──────────────────────────────────────────────────────
 
-
-@dataclass
-class VmInfo:
-    vmid: int
-    name: str
-    status: str  # "running" | "stopped"
-    config_raw: str
-
-
-def fetch_vm_info(vmid: int, adapter: ProxmoxAdapter | None = None) -> VmInfo | None:
-    if adapter is None:
-        from .services.proxmox_service import get_proxmox_adapter
-        adapter = get_proxmox_adapter()
-    runtime = adapter
-    status_result = runtime.run(["qm", "status", str(vmid)])
-    if not status_result.ok:
-        return None
-    # Parse status line like "status: running" or "status: stopped"
-    status = "stopped"
-    for line in status_result.output.splitlines():
-        if "running" in line.lower():
-            status = "running"
-            break
-    config_result = runtime.run(["qm", "config", str(vmid)])
-    config_raw = config_result.output if config_result.ok else ""
-    # Parse name from config
-    name = ""
-    for line in config_raw.splitlines():
-        if line.startswith("name:"):
-            name = line.split(":", 1)[1].strip()
-            break
-    return VmInfo(vmid=vmid, name=name, status=status, config_raw=config_raw)
+from .services.detection_service import VmInfo, fetch_vm_info  # noqa: E402
 
 
 def build_destroy_plan(vmid: int, purge: bool = False) -> list[PlanStep]:
