+++ src/osx_proxmox_next/services/__init__.py
@@ -1,5 +1,5 @@
 from . import detection_service
-from .detection_service import detect_storage_targets, detect_next_vmid, list_macos_vms
+from .detection_service import detect_storage_targets, detect_next_vmid, list_macos_vms, VmInfo, fetch_vm_info
 from .download_service import run_download_worker, check_assets
 from .preflight_service import run_preflight_worker
 from .install_service import run_dry_apply, run_live_install
@@ -11,6 +11,8 @@ __all__ = [
     "detect_storage_targets",
     "detect_next_vmid",
     "list_macos_vms",
+    "VmInfo",
+    "fetch_vm_info",
     "run_download_worker",
     "check_assets",
     "run_preflight_worker",
